Remove macro_real_block and friends, and delete text based on SCBlocks instead of runs
macro_real_block was a pain because it meant we needed to look somewhere
else every time we needed the SCBlock for something.  The new way is to
have two SCBlocks for each run, one which is the block from the flow of
the actual document (i.e. for blocks within macros, this is the
macro_real_block, otherwise it's the same as before).  The second block
("rscblock") is whichever run actually contains the text for the run.
It might be from the document, within the macro definition or within the
macro's "\contents".