Remove text run SCBlock offset
The \newpara change (commit 7f2d0abd) removed the only situation when the
start of a run might not coincide with the start of an SCBlock, or vice-versa
(although not all SCBlocks correspond to text runs, obviously).

Therefore, the offset is always zero and can be removed, simplifying the code.