Get rid of Presentation structure
It doesn't have any reason to exist.  A Narrative defines the top-level
object already.  This is, after all, a narrative-based presentation
system.

This removes a lot of faff surrounding retrieving the Narrative from the
Presentation.  It also removes a gigantic place for bugs to hide in
keeping the Presentation's view of the slides in sync with the
Narrative's opinion.

What was that quote?  "Code is terrible, and we want as little of it as
possible in our program."