Connect menu bar undo/redo/cut/copy/paste etc
It would be better to associate the menu items directly with the actions
of the GtkTextView, which would mean the menu items could be disabled
under the appropriate conditions (e.g. empty selection or nothing to
undo) without extra code.  However, there does not seem to be a way to
do that.