Re-check monitor list after a short delay
At the time the signal handler for the monitor list gets called, the new
monitor name seems to be NULL.  I suppose it needs longer to retrieve
the monitor's name.

This commit adds a 1-second timeout, after which the list is updated
again.  This fixes the problem in my tests.